+++ src/apis/templeInfo/axios.ts
@@ -4,7 +4,6 @@ export const getTempleDetails = async (templestayId: string, userId?: string) =>
   const res = await instance.get('/templestay', {
     params: { templestayId, userId },
   });
-
   return res.data;
 };
