+++ src/apis/templeInfo/prefetch.ts
@@ -1,12 +1,16 @@
 import { queryOptions } from '@tanstack/react-query';
 
-import { getTempleImages } from './axios';
-import { TemplestayImgsResponse } from './type';
+import { getTempleImages, getTempleReviews } from './axios';
+import { TemplestayImgsResponse, ReviewsResponse } from './type';
 
-const templeImagesQueryOptions = (templestayId: string) =>
+export const templeImagesQueryOptions = (templestayId: string) =>
   queryOptions<TemplestayImgsResponse>({
     queryKey: ['images', templestayId],
     queryFn: () => getTempleImages(templestayId),
   });
 
-export default templeImagesQueryOptions;
+export const templeReviewsQueryOptions = (templestayId: string, page: number) =>
+  queryOptions<ReviewsResponse>({
+    queryKey: ['reviews', templestayId, page],
+    queryFn: () => getTempleReviews(templestayId, page),
+  });
