+++ src/app/detail/[templestayId]/page.tsx
@@ -1,5 +1,5 @@
 'use client';
-import useGetTempleDetails from '@apis/templeDetail';
+import { useGetTempleDetails } from '@apis/templeInfo';
 import { useAddWishlist, useRemoveWishlist } from '@apis/wish';
 import DetailCarousel from '@components/carousel/detailCarousel/DetailCarousel';
 import ButtonBar from '@components/common/button/buttonBar/ButtonBar';
